+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## neptune-xgboost 0.9.13
 
 ### Fixes
-- Support `learning reate` for `grow_gpu_hist` updater ([#8](https://github.com/neptune-ai/neptune-xgboost/pull/8))
+- Support `learning reate` for  `updater` types ([#8](https://github.com/neptune-ai/neptune-xgboost/pull/8))
 
 
 ## neptune-xgboost 0.9.12